Use relocated model classes and update ActorRepository
Refactor imports to use org.tvl.tvlooker.domain.model for Item, User and Interaction across multiple classes following a model package change (ScoredItem, HybridRecommendationEngine, RecommendationEngine, RecommendationContext, PopularityStrategy, RecommendationStrategy). Update ActorRepository to reference ActorEntity in the JpaRepository signature and the findByTmdbId return type. These adjustments align code with the renamed/relocated model classes to restore compilation.
