+++ src/main/java/org/tvl/tvlooker/domain/motor/HybridRecommendationEngine.java
@@ -6,10 +6,10 @@ import org.tvl.tvlooker.domain.data_structure.ScoredItem;
 import org.tvl.tvlooker.domain.exception.InsufficientDataException;
 import org.tvl.tvlooker.domain.exception.InvalidEngineConfigurationException;
 import org.tvl.tvlooker.domain.exception.NoRecommendationsAvailableException;
-import org.tvl.tvlooker.domain.model.entity.Item;
+import org.tvl.tvlooker.domain.model.Item;
 import org.tvl.tvlooker.domain.motor.utils.DataProvider;
 import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
-import org.tvl.tvlooker.domain.model.entity.User;
+import org.tvl.tvlooker.domain.model.User;
 import org.tvl.tvlooker.domain.strategy.aggregation.AggregationStrategy;
 import org.tvl.tvlooker.domain.strategy.recommendation.RecommendationStrategy;
 import org.slf4j.Logger;
