+++ src/main/java/org/tvl/tvlooker/domain/strategy/recommendation/PopularityStrategy.java
@@ -1,8 +1,8 @@
 package org.tvl.tvlooker.domain.strategy.recommendation;
 
 import org.tvl.tvlooker.domain.data_structure.ScoredItem;
-import org.tvl.tvlooker.domain.model.entity.Item;
-import org.tvl.tvlooker.domain.model.entity.User;
+import org.tvl.tvlooker.domain.model.Item;
+import org.tvl.tvlooker.domain.model.User;
 import org.tvl.tvlooker.domain.motor.utils.RecommendationContext;
 
 import java.util.Comparator;
