+++ src/main/java/org/tvl/tvlooker/persistence/repository/ActorRepository.java
@@ -2,7 +2,7 @@ package org.tvl.tvlooker.persistence.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.tvl.tvlooker.domain.model.entity.Actor;
+import org.tvl.tvlooker.domain.model.entity.ActorEntity;
 
 import java.util.Optional;
 
@@ -10,12 +10,12 @@ import java.util.Optional;
  * Repository for Actor entity persistence operations.
  */
 @Repository
-public interface ActorRepository extends JpaRepository<Actor, Long> {
+public interface ActorRepository extends JpaRepository<ActorEntity, Long> {
 
     /**
      * Finds an actor by their TMDB ID.
      */
-    Optional<Actor> findByTmdbId(Long tmdbId);
+    Optional<ActorEntity> findByTmdbId(Long tmdbId);
 }
